% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frequency.R
\name{check_frequency}
\alias{check_frequency}
\title{check_frequency function}
\usage{
check_frequency(hours_frequency, time_zone)
}
\arguments{
\item{hours_frequency}{The vector containing the hours of operations.}

\item{time_zone}{The time zone.}
}
\value{
A logical vector TRUE if the EA can compute the sentiment.
}
\description{
This functions checks if the EA can send order to the plataform trading.
}
\examples{
time_zone <- "Brazil/East"
hour_freq <- generate_trade_frequency(9,17,10)
check_freq <- check_frequency(hours_frequency = hour_freq,
                             time_zone = time_zone)
}
