% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Close_Position.R
\name{Close_Position}
\alias{Close_Position}
\title{Close_Position}
\usage{
Close_Position(actual_decision)
}
\arguments{
\item{actual_decision}{The current position status("BUY IT NOW","SELL IT NOW", "SELL IT NOW CLOSE", "BUY IT NOW CLOSE" ).}
}
\value{
A vector with the new decision.
}
\description{
This functions closes a open position.
}
\examples{
decision <- 'SELL IT NOW'
decision <- Close_Position(actual_decision = decision)



}
