% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trade_frequency.R
\name{generate_trade_frequency}
\alias{generate_trade_frequency}
\title{generate_trade_frequency function}
\usage{
generate_trade_frequency(initial_time, final_time, freq_trade)
}
\arguments{
\item{initial_time}{The time the algorithm starts trading.}

\item{final_time}{The time the algorithm ends trading.}

\item{freq_trade}{The frequency which the algorithm recalculates the sentiment index.}
}
\value{
A vector containing the hours of operation.
}
\description{
generate_trade_frequency function
}
\examples{
hours_candle_10 <- generate_trade_frequency(9,17,10)
#For example, for 17:30, you should use minutes/60, i.e. 17.5
hours_candle_20 <- generate_trade_frequency(9,17.5,10)
}
