% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment_stocktwits.R
\name{get_sentiment_stocktwits}
\alias{get_sentiment_stocktwits}
\title{get_sentiment_stocktwits}
\usage{
get_sentiment_stocktwits(stock_symbol, path_twits, sentiment_index_type)
}
\arguments{
\item{stock_symbol}{A vector with the stocks symbols.}

\item{path_twits}{The path where the Json files will be stored.}

\item{sentiment_index_type}{The sentiment type to be used according to the dictionary, positive, negative or both. Default is both, positive and negative}
}
\value{
A numeric value with the value of the sentiment index.
}
\description{
This function computes the sentiment based on bullish and bearish tag from stocktwits using the last 30 twits.
}
\examples{
\dontrun{
#Not run:
path_twits <- 'your path'
  symbols <- c("EWZ", "SPX", "SPY", "USO")

 stocktwits_index <- get_sentiment_stocktwits(stock_symbol = symbols,
 path_twits = path_twits)
   }

}
