% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/havingIP.R
\name{havingIP}
\alias{havingIP}
\title{havingIP Function}
\usage{
havingIP(operational_system)
}
\arguments{
\item{operational_system}{The operational system.}
}
\description{
Function to test if the internet connection is available
}
\examples{
internet <- havingIP()

}
