% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dictionary.R
\docType{data}
\name{my_dictionary}
\alias{my_dictionary}
\title{my_dictionary}
\format{
A list with 2 components.

\describe{
\item{positive_terms}{The positive words.}
\item{negative_terms}{The negative words.}
}
}
\usage{
my_dictionary
}
\description{
A simple \code{list} containing a dictionary with positive and negative words(English and Portuguese).
}
\keyword{datasets}
