% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_hours.R
\name{operation_hours}
\alias{operation_hours}
\title{operation_hours}
\usage{
operation_hours(start_time, end_time, time_zone)
}
\arguments{
\item{start_time}{The time that the EA should start to trade.}

\item{end_time}{The time that the EA should stop to trade and close the open positions.}

\item{time_zone}{The time zone.}
}
\value{
A logical variable TRUE if the Expert Advisor can trade.
}
\description{
This function defines the operations hours of the EA.
}
\examples{
time_zone <- "Brazil/East"
op_hours<- operation_hours(start_time = 9.5,
end_time = 17,
time_zone = time_zone)





}
