% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/havingIP.R
\name{havingIP}
\alias{havingIP}
\title{havingIP Function}
\usage{
havingIP(operational_system)
}
\arguments{
\item{operational_system}{The operational system.}
}
\value{
A logical vector TRUE if internet connection is available.
}
\description{
Function to test if the internet connection is available
}
\examples{
\dontrun{
internet <- havingIP()
        }
}
